+++ src/services/TranslateService.php
@@ -231,6 +231,11 @@ class TranslateService extends Component
             }
         }
 
+        if (get_class($field) == 'benf\neo\Field' && $field->translationMethod == 'all') {
+            // special case to avoid neo overwriting blocks in all languages
+            return ['blocks' => $serialized];
+        }
+
         return $serialized;
     }